pyproject.toml
[tool.poetry]
name = "strava"
version = "0.1.0"
description = "Strava CSV exporter"
authors = ["Gerhard le Roux <gerhard.leroux@vastech.co.za>"]
readme = "README.md"
packages = [{ include = "strava" }]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32"
python-dotenv = "^1.2.2"
pandas = "^2.0"
astral = "^3.2"
timezonefinder = "^6.0"
openpyxl = "^3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.poetry.scripts]
strava-export = "strava.cli:main"
strava-sailing-logbook = "strava.cli:sailing_logbook"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
